reorg(quality-control): consolidate review infra into type-agnostic runner/validator
Create quality-control/check-report.py (type-agnostic validator loads
schemas from reviews/<type>/schema.json) and quality-control/run-review.py
(type-agnostic runner based on brooks-lint-run.py). Move templates into
reviews/{brooks,slop}/template.md. Delete stale copies at root,
.github/workflows/, and .agents/scripts/.

Also add anti-slop skill rule #21 (No hypothetical-path code) and
consolidate git-guidelines/github-* skills. Update CI workflows,
Jules reference doc, and ci-sweep-protocol to new paths.
