--- quality-control/brooks-lint-opencode.yml
@@ -1,69 +1,0 @@
-# Brooks-Lint: code review via opencode run (free model, no API key)
-name: Brooks-Lint PR Review (OpenCode)
-on:
-  pull_request:
-    types: [opened, synchronize, reopened]
-
-jobs:
-  brooks-lint:
-    runs-on: ubuntu-latest
-    permissions:
-      pull-requests: write
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Checkout brooks-lint skills
-        uses: actions/checkout@v4
-        with:
-          repository: hyhmrright/brooks-lint
-          path: .brooks-lint
-
-      - name: Setup Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.11'
-
-      - name: Install OpenCode
-        run: npm install -g opencode-ai
-
-      - name: Run Brooks-Lint via OpenCode
-        id: review
-        shell: bash
-        run: |
-          python3 .github/workflows/brooks-lint-run.py \
-            --mode review \
-            --skills-dir .brooks-lint/skills \
-            --template .github/workflows/brooks-review-template.md \
-            --base-ref "${{ github.base_ref }}" \
-            > .brooks-result.json
-
-          echo "score=$(jq -r '.score // 0' .brooks-result.json)" >> "$GITHUB_OUTPUT"
-          echo "report<<EOF" >> "$GITHUB_OUTPUT"
-          jq -r '.report // "error"' .brooks-result.json >> "$GITHUB_OUTPUT"
-          echo "EOF" >> "$GITHUB_OUTPUT"
-
-      - name: Post PR Comment
-        if: github.event_name == 'pull_request'
-        uses: actions/github-script@v7
-        with:
-          script: |
-            const fs = require('fs');
-            const r = JSON.parse(fs.readFileSync('.brooks-result.json', 'utf8'));
-            const body = `## Brooks-Lint Review (OpenCode)\n\n${r.report || "*error*"}\n\n---\n_Health Score: ${r.score ?? "N/A"}_`;
-            await github.rest.issues.createComment({
-              ...context.repo,
-              issue_number: context.issue.number,
-              body
-            });
-
-      - name: Check Threshold
-        if: ${{ vars.BROOKS_FAIL_BELOW != '' }}
-        run: |
-          threshold="${{ vars.BROOKS_FAIL_BELOW }}"
-          score=$(jq -r '.score // 0' .brooks-result.json)
-          if [ "$score" -lt "$threshold" ] 2>/dev/null; then
-            echo "Health Score $score is below threshold $threshold"
-            exit 1
-          fi
