+++ quality-control/reviews/brooks/schema.json
@@ -1,0 +1,16 @@
+{
+  "finding_required_fields": [
+    "tier",
+    "label",
+    "category",
+    "location",
+    "symptom",
+    "source",
+    "consequence",
+    "evidence"
+  ],
+  "report_required_markers": ["Finding\\s+\\d+", "Symptom:", "Source:", "Consequence:"],
+  "forbidden_categories": ["infra", "infrastructure", "ci", "workflow", "config"],
+  "report_min_length": 200,
+  "report_reject_patterns": ["-O", "optimized mode"]
+}
