+++ quality-control/reviews/slop/template.md
@@ -23,6 +23,7 @@ Baseline (from loaded skills):
 - Every assertion in a test must genuinely increase proof burden
 - Every finding must cite file paths, line numbers, and exploration evidence
 - **PEP 723 Mandate**: Any agent-authored or modified Python script that imports third-party packages MUST declare dependencies via PEP 723 inline script metadata. Reject any finding that suggests adding to `pyproject.toml` for standalone scripts.
+- **Policy Cross-Referencing**: Before reporting any finding, you MUST cross-reference it against the loaded policy documents (e.g., `bridge-burning-red-flags.md`, `runtime-control-flow-red-flags.md`). If a finding contradicts explicit policy (like the Python `-O` mode rule), it MUST be dropped. Do not report findings based on general software engineering intuition if they violate the repository's bespoke software policies.
 - **Ignore Python `-O` Mode**: We do NOT care about Python's optimized mode (`-O`) that strips `assert` statements. It is a trivial, esoteric concern. Do NOT report the use of `assert` as a problem. Findings mentioning `-O` or "optimized mode" will be rejected.
 
 ## Task: Full Repository Slop Audit
@@ -121,5 +122,5 @@ The JSON must conform to the following schema precisely:
 The ONLY way to submit your candidate report is to write the JSON to a file in the candidates directory: `{{CANDIDATES_DIR}}`.
 
 1. Write your full JSON report to a file like `{{CANDIDATES_DIR}}/candidate.json`.
-2. Do NOT try to write `.brooks-report-artifact.json` directly. The harness will validate your candidate and write the artifact itself if validation passes.
+2. Do NOT try to write `.review-report-artifact.json` directly. The harness will validate your candidate and write the artifact itself if validation passes.
 3. If the harness rejects your candidate, it will automatically restart you with a continuation prompt containing the exact validation errors.
